Rename log package files
